+++ lib/grape/exceptions/missing_group_type.rb
@@ -2,10 +2,12 @@
 
 module Grape
   module Exceptions
-    class MissingGroupTypeError < Base
+    class MissingGroupType < Base
       def initialize
         super(message: compose_message(:missing_group_type))
       end
     end
   end
 end
+
+Grape::Exceptions::MissingGroupTypeError = Grape::Exceptions::MissingGroupType