+++ lib/grape/exceptions/unsupported_group_type.rb
@@ -2,10 +2,12 @@
 
 module Grape
   module Exceptions
-    class UnsupportedGroupTypeError < Base
+    class UnsupportedGroupType < Base
       def initialize
         super(message: compose_message(:unsupported_group_type))
       end
     end
   end
 end
+
+Grape::Exceptions::UnsupportedGroupTypeError = Grape::Exceptions::UnsupportedGroupType