+++ spec/grape/exceptions/missing_group_type_spec.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+RSpec.describe Grape::Exceptions::MissingGroupType do
+  describe '#message' do
+    subject { described_class.new.message }
+
+    it { is_expected.to include 'group type is required' }
+  end
+
+  describe '#alias' do
+    subject { described_class }
+
+    it { is_expected.to eq(Grape::Exceptions::MissingGroupTypeError) }
+  end
+end