+++ spec/grape/exceptions/unsupported_group_type_spec.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+RSpec.describe Grape::Exceptions::UnsupportedGroupType do
+  subject { described_class.new }
+
+  describe '#message' do
+    subject { described_class.new.message }
+
+    it { is_expected.to include 'group type must be Array, Hash, JSON or Array[JSON]' }
+  end
+
+  describe '#alias' do
+    subject { described_class }
+
+    it { is_expected.to eq(Grape::Exceptions::UnsupportedGroupTypeError) }
+  end
+end